kms/requirements.txt
flask==0.12.2
pycryptodomex==3.6.5
